+++ pulse_lib/qblox/pulsar_uploader.py
@@ -164,7 +164,9 @@ class PulsarUploader:
         return 1e9
 
 
-    def create_job(self, sequence, index, seq_id, n_rep, sample_rate, neutralize=True):
+    def create_job(self, sequence, index, seq_id, n_rep, sample_rate,
+                   neutralize=True, alignment=None):
+        # TODO @@@ implement alignment
         # remove any old job with same sequencer and index
         self.release_memory(seq_id, index)
         return Job(self.jobs, sequence, index, seq_id, n_rep, sample_rate, neutralize)